+++ BC_training_source.py
@@ -1,24 +1,12 @@
 import gymnasium as gym
 import torch
-import random
-import numpy as np
 import argparse
 import os
 import wandb
 from sac_v2 import PolicyNetwork
-from MPC_DM_model import ReplayBuffer, MPC_DM
-from MPC_v2 import ReplayBuffer_traj, MPC
-from stable_baselines3 import SAC
-
-def seed_everything(seed):
-    random.seed(seed)
-    np.random.seed(seed)
-    os.environ['PYTHONHASHSEED'] = str(seed)
-    torch.manual_seed(seed)
-    torch.cuda.manual_seed(seed)
-    torch.cuda.manual_seed_all(seed)
-    torch.backends.cudnn.deterministic = True
-    torch.backends.cudnn.benchmark = False  
+from MPC_DM_model import MPC_DM
+#from stable_baselines3 import SAC
+from utils import seed_everything, load_buffer
 
 parser = argparse.ArgumentParser()
 parser.add_argument("MPC_pre_ep", type=int, nargs='?', default=2000)
@@ -33,21 +21,15 @@ seed_everything(args.seed)
 
 if args.env == "reacher":
     source_env = "Reacher-v4"
-    target_env = "Reacher-3joints"
     source_s_dim = 11
     source_a_dim = 2
-    target_s_dim = 14
-    target_a_dim = 3
 elif args.env == "cheetah":
     source_env = "HalfCheetah-v4"
-    target_env = "HalfCheetah-3legs"
     source_s_dim = 18
     source_a_dim = 6
-    target_s_dim = 23
-    target_a_dim = 9
 
 if args.wandb:
-    wandb.init(project="cdpc", name = f'BC training', tags=["cdpc"])
+    wandb.init(project="cdpc", name = f'BC training source', tags=["policy&DM"])
 location = f'./models/{args.env}/seed_{str(args.seed)}/'
 mpc_location = f'{location}/expert_ratio_{args.expert_ratio}/'
 hidden_dim = 256
@@ -55,29 +37,33 @@ action_range = 10.0 if args.env=="reacher" else 1.0
 
 # agent_target = SAC.load('models/cheetah/seed_7/HalfCheetah-3legs_SAC_3_128_200000_2.zip', device=args.device) # SB3
 agent = PolicyNetwork(source_s_dim, source_a_dim, hidden_dim, action_range, args.device).to(args.device)
-# print( f'{location}{str(args.seed)}_{args.env}_source.pth' )
 agent.load_state_dict(torch.load( f'{location}{str(args.seed)}_{args.env}_source.pth', weights_only=True, map_location=args.device ))
 agent_medium = PolicyNetwork(source_s_dim, source_a_dim, hidden_dim, action_range, args.device).to(args.device)
 agent_medium.load_state_dict(torch.load( f'{location}{str(args.seed)}_{args.env}_source_medium.pth', weights_only=True, map_location=args.device ))
 
 source_buffer_path = f'./train_set/{str(args.seed)}_{args.env}_{args.expert_ratio}_source_buffer_expert.pkl'
-
-import pickle
-with open(source_buffer_path, 'rb') as f:
-    buffer = pickle.load(f)
+buffer = load_buffer(source_buffer_path)
 
 mpc_dm = MPC_DM(source_s_dim, source_a_dim, args.device)
 
-batch_size=512
-print('start training...')
+if not os.path.exists(f'{mpc_location}/{str(args.seed)}_DynamicModel_source.pth'):
+    print("##### Source Domain Dynamic Model #####")
+    os.makedirs(mpc_location, exist_ok=True)
+    batch_size=512
 
-for i in range(args.MPC_pre_ep):
-    loss_mpc, loss_dm = mpc_dm.update(batch_size, buffer)
-    if args.wandb:
-        wandb.log({"mpc_dm episode": i, "train/loss_mpc": loss_mpc, "train/loss_dm": loss_dm,})
+    for i in range(args.MPC_pre_ep):
+        loss_mpc, loss_dm = mpc_dm.update(batch_size, buffer)
 
-os.makedirs(mpc_location, exist_ok=True)
-torch.save(mpc_dm.mpc_policy_net.state_dict(), f'{mpc_location}/{str(args.seed)}_MPCModel_source.pth')
-torch.save(mpc_dm.dynamic_model.state_dict(), f'{mpc_location}/{str(args.seed)}_DynamicModel_source.pth')
+        if args.wandb:
+            wandb.log({"mpc_dm episode": i, "train/loss_mpc": loss_mpc, "train/loss_dm": loss_dm,})
 
-print(f'{mpc_location}/{str(args.seed)}_DynamicModel_source.pth saved')
+        if i % 10==0:
+            env = gym.make(source_env)
+            reward = mpc_dm.evaluate_policy(env, args.seed)
+            print(f"episode: {i}/{args.MPC_pre_ep}, reward: {reward}")
+            if args.wandb:
+                wandb.log({"mpc_dm episode": i, "test/BC_score": reward,})
+
+    #torch.save(mpc_dm.mpc_policy_net.state_dict(), f'{mpc_location}/{str(args.seed)}_MPCModel_source.pth')
+    torch.save(mpc_dm.dynamic_model.state_dict(), f'{mpc_location}/{str(args.seed)}_DynamicModel_source.pth')
+    print(f'{mpc_location}/{str(args.seed)}_DynamicModel_source.pth saved')
